Separate generated from non-generated code.
By merging the generated interface and the runtime specific code and
turning Builder into a parameterized trait, we can avoid violating
the orphan rule. This enables moving a lot of stuff to the
non-generated part of the code, only keeping what really needs to be
generated.
