+++ Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
-members = ["com-api", "com-api-example", "com-api-sample-instance", "com-api-sample-interface", "com-api-sample-runtime"]
+members = ["com-api", "com-api-example", "com-api-sample-gen", "com-api-sample-runtime"]
 resolver = "3"
