+++ com-api-example/Cargo.toml
@@ -9,6 +9,5 @@ license = "Apache-2.0"
 
 [dev-dependencies]
 com-api = { path = "../com-api" }
-com-api-sample-interface = { path = "../com-api-sample-interface" }
-com-api-sample-instance = { path = "../com-api-sample-instance" }
-futures = "0.3"
+com-api-sample-gen = { path = "../com-api-sample-gen" }
+com-api-sample-runtime = { path = "../com-api-sample-runtime" }
