--- com-api-sample-instance/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "com-api-sample-instance"
-version = "0.1.0"
-edition = "2024"
-publish = ["common"]
-
-[dependencies]
-com-api-sample-runtime = { path = "../com-api-sample-runtime" }
-com-api-sample-interface = { path = "../com-api-sample-interface" }
-com-api = { path = "../com-api" }
